+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 #2016.11.10
 ##Add "PlotEff.py", bug fixed
 ##Add "CHANGLOG.md"
+
+#2016.11.11
+##Add mass point label in legend.
+##Add "DrawAllFile" Function